Keep track of i offset between different buffers
This commit fixes a bug that occurs on inputs greated than BUFFER_SIZE when the key length is not a divisor of the buffer length. As we start again on the next buffer, the offset for xor-ing is reset to 0, but that also affects the offset for the key index, which shoul be retained in between buffers.
We now keep track of the last value we were at on the last buffer and apply that offset to our key reads on the next one. This fixes the bug.
While I'm at it I also moved the key length calculation outside of the loop, which was stupid.